fixes for url

+++ src/pages/api/profile/image/[id].ts
@@ -54,21 +54,21 @@ async function create(
                     const imagePath = path.resolve(`./public/uploads/profile/${id}/${fileData.originalFilename}`)
                     console.log("moving file: ", fileData.filepath, " to ", imagePath)
                     fs.renameSync(fileData.filepath, imagePath)
-                    fs.unlink(fileData.filepath, (deleteError) => {
+                    if (fs.existsSync(fileData.filepath)) fs.unlink(fileData.filepath, (deleteError) => {
                         if (deleteError) console.log("Temporary file delete error", deleteError)
                         console.log('Temporary file deleted!')
                     })
                     const profileData: Profile | null = await prisma.profile.findUnique({ where: { id: id } })
                     if (profileData && profileData.imageUrl) {
-                        const oldImagePath = path.resolve(`./public${profileData.imageUrl}`)
-                        fs.unlink(oldImagePath, (deleteError) => {
+                        const oldImagePath = path.resolve(`./public/${profileData.imageUrl}`)
+                        if (fs.existsSync(oldImagePath)) fs.unlink(oldImagePath, (deleteError) => {
                             if (deleteError) console.log("Old file delete error", deleteError)
                             console.log('Old file deleted!')
                         })
                     }
                     await prisma.profile.update({
                         where: { id: id },
-                        data: { imageUrl: `/uploads/profile/${id}/${fileData.originalFilename}` }
+                        data: { imageUrl: `uploads/profile/${id}/${fileData.originalFilename}` }
                     })
                     res.status(200).json({ data: { filed: fields, file: files } })
                 })
